Added Column Major Initialization
Modified the program to add column major initialization through
initializer lists. To distinguish between row and column major
assignment, the user has to enter an empty vector at the end of
initializer list to signify that it is a column major assignment.

The main.cpp file was also updated to check whether it is working
properly, so some matrices have been initialized by column using this
method.

The actual storage behind the scenes is row major only.